requirements.txt
absl-py==0.13.0
aio-pika==6.8.2
aiofiles==22.1.0
aiohttp==3.7.4
aiormq==3.3.1
APScheduler==3.7.0
aspose-words==22.6.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
autopep8==2.0.0
beautifulsoup4==4.11.1
bidict==0.22.0
boto3==1.26.39
botocore==1.29.39
CacheControl==0.12.11
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
cloudpickle==1.6.0
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.0.4
cryptography==38.0.4
cycler==0.11.0
dask==2021.11.2
dnspython==1.16.0
docopt==0.6.2
et-xmlfile==1.1.0
fbmessenger==6.0.0
fire==0.5.0
flatbuffers==2.0.7
fsspec==2022.11.0
future==0.18.2
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.51.1
h5py==3.7.0
htmldocx==0.0.6
httptools==0.5.0
humanfriendly==10.0
idna==3.3
importlib-metadata==5.2.0
jmespath==1.0.1
joblib==1.0.1
jsonpickle==2.0.0
jsonschema==3.2.0
kafka-python==2.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
locket==1.0.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.3.4
mattermostwrapper==2.2
msgpack==1.0.4
multidict==5.2.0
networkx==2.6.3
numpy==1.19.5
oauthlib==3.2.2
openai==0.25.0
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==20.9
pamqp==2.3.0
pandas==1.3.4
pandas-stubs==1.5.2.221124
partd==1.3.0
Pillow==9.3.0
prompt-toolkit==2.0.10
protobuf==3.19.6
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydot==1.4.2
PyJWT==2.6.0
pykwalify==1.8.0
pymongo==3.10.1
pyngrok==5.1.0
pyparsing==3.0.9
pyreadline3==3.4.1
pyrsistent==0.19.3
pyTelegramBotAPI==3.8.3
python-crfsuite==0.9.8
python-dateutil==2.8.2
python-docx==0.8.11
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2020.1
PyYAML==6.0
pyzmq==22.3.0
questionary==1.10.0
randomname==0.1.5
rasa==3.1.0
rasa-sdk==3.1.3
redis==3.5.3
regex==2021.8.28
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rocketchat-API==1.16.0
rsa==4.9
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==0.24.2
scipy==1.9.3
scoop==0.7.1.1
sentry-sdk==1.3.1
six==1.16.0
sklearn-crfsuite==0.3.6
slackclient==2.9.4
soupsieve==2.3.2.post1
SQLAlchemy==1.4.45
tabulate==0.9.0
tarsafe==0.0.3
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.4
tensorflow-addons==0.15.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.1.1
terminaltables==3.1.10
threadpoolctl==3.1.0
tomli==2.0.1
toolz==0.12.0
tqdm==4.64.1
twilio==6.50.1
typeguard==2.13.3
types-pytz==2022.6.0.1
typing-extensions==3.10.0.2
typing-utils==0.1.0
tzlocal==2.1
ujson==4.3.0
urllib3==1.26.9
wcwidth==0.2.5
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.2
zipp==3.11.0
